fix: dconfig类型默认为float64，直接转化成uint32会导致崩溃
先转成float64再转uint32
本地升级后，出现dde-session-daemon崩溃

Log: 类型转化
Influence: dde-session-daemon崩溃

+++ network/manager.go
@@ -248,7 +248,7 @@ func (m *Manager) init() {
 					logger.Warning(err)
 					return
 				}
-				m.resetWifiOSDEnableTimeout = v.Value().(uint32)
+				m.resetWifiOSDEnableTimeout = uint32(v.Value().(float64))
 			}
 
 			getProtalAuthEnable()
